Extracted the user attr handling methods from ConfigModelV7 into its own class
